Remove favicon ico from root of repo
Even w the favicon ico removed from the root of the repo, all three
pages, the Flammarion, the Content page, and the index md render the
favicon. There is a favicon ico in the root of the docs folder from
which the pages are served. But, remotely … the index dot md page again
fails to render the corresponding favicon. Therefore, am adding favicon
 ico to root of main url … rwebaz.gthub.io.